Added duration reporting to compact reporter
Also made the duration formatting code available to all reporters.

Closes #780
